be-predictor/backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.31.1
python-dotenv==1.0.1
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
joblib==1.4.2
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

